+++ src/test/groovy/org/codenarc/rule/unnecessary/UnnecessarySetterRuleTest.groovy
@@ -35,7 +35,9 @@ class UnnecessarySetterRuleTest extends AbstractRuleTestCase<UnnecessarySetterRu
             x.set(1)
             x.setup(2)
             x.setURL('')
-            x.setSomething('arg1', 'arg2') '''
+            x.setSomething('arg1', 'arg2')
+            x.setE(8, 9, 10)
+        '''
         assertNoViolations(SOURCE)
     }
 
@@ -53,15 +55,28 @@ class UnnecessarySetterRuleTest extends AbstractRuleTestCase<UnnecessarySetterRu
             x.setProperty(1)
             x.setProperty(this.getA())
             x.setProperty([])
+            x.setE(3)           // see #364
         '''
         assertViolations(SOURCE,
             [lineNumber:2, sourceLineText:'x.setProperty(1)', messageText:'setProperty(..) can probably be rewritten as property = ..'],
             [lineNumber:3, sourceLineText:'x.setProperty(this.getA())', messageText:'setProperty(..) can probably be rewritten as property = ..'],
-            [lineNumber:4, sourceLineText:'x.setProperty([])', messageText:'setProperty(..) can probably be rewritten as property = ..'])
+            [lineNumber:4, sourceLineText:'x.setProperty([])', messageText:'setProperty(..) can probably be rewritten as property = ..'],
+            [lineNumber:5, sourceLineText:'x.setE(3)', messageText:'setE(..)'])
     }
 
     @Test
-    void callingSettersOnSuperDoesNotCauseViolations() {
+    void testCallingSetterWithinExpression_NoViolations() {
+        final SOURCE = '''
+            if (!file.setExecutable(true)) {
+                throw new Exception("Cannot set ${file} as executable")
+            }
+            def count = x.setCount(92)
+        '''
+        assertNoViolations(SOURCE)
+    }
+
+    @Test
+    void testCallingSettersOnSuperDoesNotCauseViolations() {
         final SOURCE = '''
             super.setProperty(1)
         '''